Nitpicks for second article
Some nitpick corrections.

I am not sure if I understood the following sentence correctly: `- The standard is written for the C programming language. That means lots of C-typical structures are present, such as linked lists in the [boot information].`

In its original form it was not really clear what is meant here. I tried to guess and correct in that way, but briefly glancing over the linked boot information, I was not able to spot linked lists :confused: