Reduce cache size on travis
The Windows build fails often because it takes too long to store the build cache (timeout). This commit uses `cargo cache` to clean up the /home/philipp/.cargo directory before caching. It also removes caching of the `target` directory since it needs to be recompiled for each new nightly anyway.