Simplify the later example error
Required so that the current compiler output match the example error on line 637, (because the compiler is now clever enough to warn about):
```bash
warning: variable does not need to be mutable
 --> src/main.rs:5:9
  |
5 |     let mut heap_value = Box::pin(SelfReferential {
  |         ----^^^^^^^^^^
  |         |
  |         help: remove this `mut`
  |
  = note: `#[warn(unused_mut)]` on by default
```
Which is distracting. (Also a chance to passively introduce `#[allow(unused_mut)]` )  

Though this proposed change may, in itself be a distraction for the student?