Add additional “real hardware” instructions
Because macOS and Linux are both Unix-like operating systems, they both have the `dd` command bundled. During my 13 years of having Linux as my daily driver desktop OS and 1 year of having a MacBook Pro on the side, I’ve used this command long enough to know that it definitely gets this job done. So, time to start contributing.