Update vga-text-buffer failed link for normal reads and write.
Previously the hyperlink for 'supports normal reads and writes' in post 3 was redirecting to a page not found, as I guess Stanford has remove that link. So I have added the correct link, which redirects to Stanford's Secure Systems Group, where they have the deails about CS140 course and relevant pintos docs.